Fix types and class structure.
