+++ src/ebfm/coupling/couplers/yacCoupler.py
@@ -6,7 +6,6 @@ import yac
 import numpy as np
 
 from ebfm.core import logging
-from ebfm.core.constants import SECONDS_PER_DAY, DAYS_PER_YEAR
 
 from .base import Coupler, Grid, GridDict, CouplingConfig, CouplerErrorCode
 
@@ -42,18 +41,6 @@ class YACCoupler(Coupler[yac.ExchangeType]):
         # will be initialized in self._add_grid()
         self.grid: yac.UnstructuredGrid = None
         self.cell_centers: yac.Points = None
-
-    def get_conversion_per_year_factor(self) -> float:
-        """
-        Returns the conversion factor to scale a per-timestep value to a per-year value.
-
-        Performs the calculation based on the current time step size stored in the instance.
-
-        @note This method assumes that self.time has been set and contains the key 'dt'.
-
-        @return The conversion factor (unitless).
-        """
-        return (DAYS_PER_YEAR * SECONDS_PER_DAY) / (self._time["dt"] * SECONDS_PER_DAY)
 
     def setup(self, grid: GridDict, time: dict[str, float]):
         """
